# use journal_mode=MEMORY for performance