--- a/lib/Database.js
+++ b/lib/Database.js
@@ -44,7 +44,7 @@ Database.prototype.configure = function(){
   this.db.pragma('page_size=4096'); // (default: 1024)
   this.db.pragma('cache_size=-2000'); // (default: -2000, 2GB)
   this.db.pragma('synchronous=OFF');
-  this.db.pragma('journal_mode=OFF');
+  this.db.pragma('journal_mode=MEMORY');
   this.db.pragma('temp_store=MEMORY');
 };