--- a/lib/DocStore.js
+++ b/lib/DocStore.js
@@ -8,6 +8,7 @@ util.inherits( DocStore, Database );
 
 DocStore.prototype.reset = function(){
   this.db.exec('DROP TABLE IF EXISTS docs');
+  this.db.exec('DROP TABLE IF EXISTS rtree');
   this.db.exec('CREATE TABLE docs( id INTEGER PRIMARY KEY, json TEXT )');
 
   // create rtree table