When using `pelias/docker` to create a placeholder database today I found that it was very sloooow.

I watched the `PLACEHOLDER_DATA` dir in another terminal and noticed that the journal file was being created despite us setting `journal_mode=OFF`.

I'm not 100% sure why this is happening 🤷 but it's fixed by setting `journal_mode=MEMORY` instead 🎉 

I noticed a 14x speed improvement as a result of using this configuration.

```bash
# before
node cmd/load.js  5.97s user 24.37s system 93% cpu 32.439 total

# after
node cmd/load.js  3.77s user 1.69s system 108% cpu 5.038 total
```