@missinglink 2020-04-24
we have the same problem for interpolation in https://github.com/pelias/interpolation/blob/a1d3a3958c4db50ff3718c39ca7c1e69513b9a13/query/configure.js#L9

I will have a look into it, maybe a `better-sqlite3` issue?