@missinglink 2020-04-24
https://github.com/JoshuaWise/better-sqlite3/pull/380